🔄 v1.0.1-alpha – Structure Overhaul Huge thanks to CJ from the modding Discord for the feedback on project structure! We've now reorganized the mod into a scalable format with proper separation of core logic and Harmony patches:
SPT-Stances/
├── Core/
│   ├── StanceController.cs
│   └── StanceInput.cs
├── Patches/
│   └── WeaponPosePatch.cs
├── config.json
├── manifest.json
└── README.md
This should make collaboration and future updates a lot cleaner 🔧
Feel free to fork, PR, or give brutal feedback—learning in public 💪
