+++ Patches/WeaponPosePatrch.cs
@@ -1,4 +1,5 @@
 ﻿// --- File: Patches/WeaponPosePatch.cs ---
+using Core;
 using HarmonyLib;
 using UnityEngine;
 
@@ -15,21 +16,21 @@ namespace SPTStances.Patches
             Vector3 stanceOffset = Vector3.zero;
             Quaternion stanceRotation = Quaternion.identity;
 
-            switch (SPTStances.StanceController.CurrentStance)
+            switch (StanceController.CurrentStance)
             {
-                case SPTStances.WeaponStance.HighReady:
+                case WeaponStance.HighReady:
                     stanceOffset = new Vector3(0f, 0.08f, -0.05f);
                     stanceRotation = Quaternion.Euler(10f, 0f, 0f);
                     break;
-                case SPTStances.WeaponStance.LowReady:
+                case WeaponStance.LowReady:
                     stanceOffset = new Vector3(0f, -0.05f, -0.02f);
                     stanceRotation = Quaternion.Euler(-5f, 0f, 0f);
                     break;
-                case SPTStances.WeaponStance.Compressed:
+                case WeaponStance.Compressed:
                     stanceOffset = new Vector3(0f, -0.1f, -0.08f);
                     stanceRotation = Quaternion.Euler(0f, 0f, 10f);
                     break;
-                case SPTStances.WeaponStance.Default:
+                case WeaponStance.Default:
                 default:
                     return;
             }
